+++ src/it/scala/au/com/agl/arc/extract/AvroExtractSuite.scala
@@ -94,14 +94,14 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
       BytesExtract(
         name="dataset",
         description=None,
-        cols=Right(Nil),
         outputView=outputView, 
         input=Right(targetBinaryFile),
         authentication=None,
         persist=false,
         numPartitions=None,
         contiguousIndex=true,
-        params=Map.empty
+        params=Map.empty,
+        failMode=FailModeTypeFailFast
       )
     )