+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 - added `AzureCosmosDBExtract` stage.
 - added ability to pass `params` to [Lifecycle Plugins](https://aglenergy.github.io/arc/extend/#lifecycle-plugins).
 - rewrote tutorial to use [arc-starter](https://github.com/seddonm1/arc-starter).
-- added `schemaView` and `schemaURI` options to `BytesExtract` to allow pipeline to continue if missing binary files.
+- added `failMode` to `BytesExtract` to allow pipeline to continue if missing binary files.
 - added `DataFramePrinterLifecyclePlugin` to base image.
+- `ARC.run()` now returns the final `Option[DataFrame]` facilitating better [integrations](https://github.com/seddonm1/arc-starter).
 
 # 1.12.2